pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "demagnetize"
dynamic = ["version"]
description = "Convert magnet links to .torrent files"
readme = "README.rst"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "John Thorvald Wodder II", email = "demagnetize@varonathe.org" }
]

keywords = [
    "bittorrent",
    "magnet link",
    "torrent",
]

classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Environment :: Console",
    "Framework :: AnyIO",
    "Topic :: Communications :: File Sharing",
    "Typing :: Typed",
]

dependencies = [
    "anyio ~= 4.0",
    "attrs >= 20.1.0",
    "click >= 8.0",
    "click-loglevel ~= 0.5",
    "colorlog ~= 6.0",
    "httpx ~= 0.22",
    "torf @ git+https://github.com/mhdzumair/torf.git",
    "yarl ~= 1.7",
]

[project.scripts]
demagnetize = "demagnetize.__main__:main"

[project.urls]
"Source Code" = "https://github.com/jwodder/demagnetize"
"Bug Tracker" = "https://github.com/jwodder/demagnetize/issues"

[tool.hatch.version]
path = "src/demagnetize/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    "/docs",
    "/src",
    "/test",
    "CHANGELOG.*",
    "CONTRIBUTORS.*",
    "tox.ini",
]

[tool.hatch.envs.default]
python = "3"

[tool.mypy]
allow_incomplete_defs = false
allow_untyped_defs = false
ignore_missing_imports = false
# <https://github.com/python/mypy/issues/7773>:
no_implicit_optional = true
implicit_reexport = false
local_partial_types = true
pretty = true
show_error_codes = true
show_traceback = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "torf.*"
ignore_missing_imports = true
